2.5update

+++ ita_root/common_libs/validate/valid_50104.py
@@ -322,7 +322,7 @@ def menu_column_valid(objdbca, objtable, option):
         if retBool and integer_maximum_value is None:
             integer_maximum_value = 2147483648
         # 最大値<最小値になっている場合、エラー
-        if retBool and int(integer_minimum_value) > int(integer_maximum_value):
+        if retBool and float(integer_minimum_value) > float(integer_maximum_value):
             retBool = False
             msg = g.appmsg.get_api_message("MSG-20068", [integer_minimum_value, integer_maximum_value])
         # 小数最小値が設定されている場合、エラー
@@ -392,7 +392,7 @@ def menu_column_valid(objdbca, objtable, option):
         # 初期値(整数) 最大数、最小数をチェック
         if retBool and integer_default_value:
             if isinstance(integer_default_value, int):
-                if int(integer_maximum_value) < int(integer_default_value) or int(integer_default_value) < int(integer_minimum_value):
+                if float(integer_maximum_value) < float(integer_default_value) or float(integer_default_value) < float(integer_minimum_value):
                     retBool = False
                     msg = g.appmsg.get_api_message("MSG-20085", [])
